+++ node/src/components/contract_runtime/types/step_outcome.rs
@@ -1,0 +1,12 @@
+use crate::types::DataSize;
+use casper_types::{execution::Effects, EraId, PublicKey, U512};
+use std::collections::BTreeMap;
+
+/// Effects from running step and the next era validators that are gathered when an era ends.
+#[derive(Clone, Debug, DataSize)]
+pub(crate) struct StepOutcome {
+    /// Validator sets for all upcoming eras that have already been determined.
+    pub(crate) upcoming_era_validators: BTreeMap<EraId, BTreeMap<PublicKey, U512>>,
+    /// An [`Effects`] created by an era ending.
+    pub(crate) step_effects: Effects,
+}
