+++ node/src/components/contract_runtime/types/validator_weights_by_era_id_request.rs
@@ -1,0 +1,42 @@
+use casper_storage::data_access_layer::EraValidatorsRequest;
+use casper_types::{Digest, EraId, ProtocolVersion};
+
+/// Request for validator weights for a specific era.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ValidatorWeightsByEraIdRequest {
+    state_hash: Digest,
+    era_id: EraId,
+    protocol_version: ProtocolVersion,
+}
+
+impl ValidatorWeightsByEraIdRequest {
+    /// Constructs a new ValidatorWeightsByEraIdRequest.
+    pub fn new(state_hash: Digest, era_id: EraId, protocol_version: ProtocolVersion) -> Self {
+        ValidatorWeightsByEraIdRequest {
+            state_hash,
+            era_id,
+            protocol_version,
+        }
+    }
+
+    /// Get the state hash.
+    pub fn state_hash(&self) -> Digest {
+        self.state_hash
+    }
+
+    /// Get the era id.
+    pub fn era_id(&self) -> EraId {
+        self.era_id
+    }
+
+    /// Get the protocol version.
+    pub fn protocol_version(&self) -> ProtocolVersion {
+        self.protocol_version
+    }
+}
+
+impl From<ValidatorWeightsByEraIdRequest> for EraValidatorsRequest {
+    fn from(input: ValidatorWeightsByEraIdRequest) -> Self {
+        EraValidatorsRequest::new(input.state_hash)
+    }
+}
